Added ObservableGenerator for seemless client binding (no need to create an ObservableXX object).
Now using GWT.create to form the binding custom object will give you an ObservableXX object.
This is working seemlessly for me since I use Gwt-Jackson with a @JsonCreator to return a deferred object.

Note that I have added some edited core GWT code, for the <when-type-annotated> condition handler.
I have submitted this code to the GWT repository, but its up to them whether or not they accept it.